refactor: router 컴포넌트명 규칙적용
